try to make git diff easier to understand by not moving files into other directory
